pyproject.toml
[project]
name = "complaint-ai"
version = "0.1.0"
description = "AI-powered consumer complaint classification and resolution pipeline"
requires-python = ">=3.11"
dependencies = [
    # Web API
    "python-dotenv>=1.0,<2",
    "fastapi>=0.110,<1",
    "uvicorn[standard]>=0.27,<1",
    # Data & persistence
    "pydantic>=2.6,<3",
    "sqlalchemy>=2.0,<3",
    "psycopg2-binary>=2.9,<3",
    "pgvector>=0.2.4,<0.4",
    # LLM stack — Google ADK + Gemini
    "google-adk>=1.0.0",
    "google-genai>=1.0.0",
    "sentence-transformers>=2.2,<4",
    "typing-extensions>=4.9",
    # Observability (OpenTelemetry + LangSmith with native ADK integration)
    "opentelemetry-api>=1.20,<2",
    "opentelemetry-sdk>=1.20,<2",
    "langsmith[google-adk]>=0.1",
    "pandas>=3.0.2",
    "python-multipart>=0.0.26",
    "pypdf>=6.10.2",
]

[project.optional-dependencies]
test = [
    "pandas>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# Use CPU-only PyTorch — no NVIDIA CUDA packages needed (no GPU on server)
[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
